+++ CHANGELOG.md
@@ -30,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   most grammars using strings directly. Includes `SegmentGenerator` objects
   to bind dialect objects at runtime from sets. Thanks [@Katzmann1983](https://github.com/Katzmann1983)!
 - Rule `L029` for using unreserved keywords as variable names.
+- The jinja templater now allows macros loaded from files, and the
+  hydration of variables ending in `_path` in the config files.
+- JSON operators and the `DISTINCT ON ()` syntax for the postgres dialect.
 
 ### Changed
 
@@ -57,6 +60,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   fixing of files, if one fix would re-introduce a problem with another.
 - Behaviour of the `Bracketed` grammar has been changed to treat it's
   content as a `Sequence` rather than a `OneOf`.
+- Move to `SandboxedEnvironment` rather than `Environment` for jinja
+  templating for security.
+- Improve reporting of templating issues, especially for the jinja templater
+  so that missing variables are rendered as blanks, but still reported as
+  templating violations.
 
 ## [0.3.1] - 2020-02-17